+++ src/Localization/TEXTS_DE_CH.mjs
@@ -1,6 +1,6 @@
 import { LOCALIZATION_KEY_LANGUAGE, LOCALIZATION_KEY_SYSTEM, LOCALIZATION_KEY_SYSTEM_WITH_LANGUAGE } from "./LOCALIZATION_KEY.mjs";
 
-export const TEXTS_DE = Object.freeze({
+export const TEXTS_DE_CH = Object.freeze({
     [LOCALIZATION_KEY_LANGUAGE]: "Sprache",
     [LOCALIZATION_KEY_SYSTEM]: "System",
     [LOCALIZATION_KEY_SYSTEM_WITH_LANGUAGE]: "System ({language})"